Move DatasourceChecker to package preparer/checker
