+++ sharding-scaling/sharding-scaling-bootstrap/src/test/java/org/apache/shardingsphere/shardingscaling/fixture/FixtureH2DataSourceChecker.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.shardingscaling.fixture;
 
-import org.apache.shardingsphere.shardingscaling.core.execute.executor.checker.DataSourceChecker;
+import org.apache.shardingsphere.shardingscaling.core.preparer.checker.DataSourceChecker;
 
 import javax.sql.DataSource;
 import java.util.Collection;
