+++ sharding-scaling/sharding-scaling-mysql/src/main/java/org/apache/shardingsphere/shardingscaling/mysql/MySQLDataSourceChecker.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.shardingscaling.mysql;
 
 import org.apache.shardingsphere.shardingscaling.core.exception.DatasourceCheckFailedException;
-import org.apache.shardingsphere.shardingscaling.core.execute.executor.checker.AbstractDataSourceChecker;
+import org.apache.shardingsphere.shardingscaling.core.preparer.checker.AbstractDataSourceChecker;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
